+++ cnccode/data_classes/DBEJCallActivity.php
@@ -70,7 +70,7 @@ class DBEJCallActivity extends DBECallActivity
     const emptyAssetReason                    = "emptyAssetReason";
     const emailSubjectSummary                 = "emailSubjectSummary";
     const prePayChargeApproved                = "prePayChargeApproved";
-
+    const automateMachineID                   = "automateMachineID";
     var $fromString;
 
     /**
@@ -388,6 +388,13 @@ class DBEJCallActivity extends DBECallActivity
             DA_NOT_NULL,
             "problem.prePayChargeApproved"
         );
+        $this->addColumn(
+            self::automateMachineID,
+            DA_INTEGER,
+            DA_NULL,
+            "problem.automateMachineID"
+        );
+        
         $this->setAddColumnsOff();
         $this->fromString = $this->getTableName(
             ) . " LEFT JOIN " . " callacttype ON caa_callacttypeno = cat_callacttypeno" . " LEFT JOIN " . " item AS activity_type_item ON cat_itemno = activity_type_item.itm_itemno" . " LEFT JOIN " . " consultant ON consultant.cns_consno = caa_consno" . " LEFT JOIN " . " problem as problem ON problem.pro_problemno = callactivity.caa_problemno" . " LEFT JOIN " . " customer ON cus_custno = pro_custno" . " LEFT JOIN " . " address ON add_custno = pro_custno" . " AND add_siteno = caa_siteno" . " LEFT JOIN " . " custitem AS contract ON problem.pro_contract_cuino = contract.cui_cuino" . " LEFT JOIN " . " item AS contractitem ON contract.cui_itemno = contractitem.itm_itemno" . " LEFT JOIN " . " project ON problem.pro_projectno = project.projectID" . " LEFT JOIN " . " contact ON con_contno = caa_contno" . " LEFT JOIN " . " consultant as completed_user ON callactivity.caa_completed_consno = completed_user.cns_consno" . " LEFT JOIN " . " rootcause ON rootcause.rtc_rootcauseno = problem.pro_rootcauseno" . " LEFT JOIN  " . " team ON consultant.teamID = team.teamID";