+++ htdocs/components/ActivityComponent/subComponents/ActivityEditComponent.js
@@ -263,7 +263,8 @@ class ActivityEditComponent extends MainComponent {
             "submitAsOvertime",
             "emptyAssetReason",
             "completeDate",
-            "Inbound"
+            "Inbound",
+            "automateMachineID"
         ]);
         this.api
             .updateActivity(finalData)
@@ -1759,7 +1760,8 @@ class ActivityEditComponent extends MainComponent {
         if (!data || !data.customerId) {
             return '';
         }
-        return <AssetListSelectorComponent
+        return <div className="flex-row">
+            <AssetListSelectorComponent
             emptyAssetReason={data.emptyAssetReason}
             assetName={data.assetName}
             assetTitle={data.assetTitle}
@@ -1768,16 +1770,22 @@ class ActivityEditComponent extends MainComponent {
             onChange={value => this.handleAssetSelect(value)}
             showUnsupportedWhileSelected={true}
         />
+            {data.automateMachineID?<i className="fal fa-cog ml-5 pointer fa-2x" 
+                            onClick={()=>window.open(`https://serverguard.cnc-ltd.co.uk/automate/computer/${data.automateMachineID}/normal-tiles`,"_target")}
+                            
+                            ></i>:null}
+            </div>
     }
     handleAssetSelect = (value) => {
         const {data} = this.state;
         data.assetName = "";
         data.assetTitle = "";
-        data.emptyAssetReason = "";
+        data.emptyAssetReason = "";        
         if (value) {
             if (value.isAsset) {
                 data.assetName = value.name;
                 data.assetTitle = value.name + " " + value.LastUsername + " " + value.BiosVer;
+                data.automateMachineID=value.ComputerID;
             } else {
                 data.emptyAssetReason = value.template;
             }